refactor: improve abstraction, improve data collection
This commit introduces a new layer of abstraction for data collection. Particularly, there is now a class meant to abstract away shell commands. There is also a new class to better represent files that changed in commits, and another to abstract how renames are saved.

In the data collection department, there are now A LOT more situations covered - particularly when it comes to recognizing files that were deleted, and properly persisting change information when a file is renamed. They attempt to be as extensive as possible, but some edge cases may have been missed despite the effort.

More documentation was also added.