feat: keep only entities in the result, save file changes
A new experiment on the analysis is introduced in this commit. We want
to see if performing decompositions using only entities yields better
results. Therefore, any non-entity files are removed from the final
json/dictionary. This new approach assumes that an IDToEntity file was
saved in a particular location. If the file does not exist, then there
is no culling of data - it's the same as the previous approach.

An intermediary step was re-added, where the file changes are stored
to the disk and retrieved on subsequent runs. This should speed up
development.