Add new more efficient collection strategy
Previous data collection was quite slow, due to repeated shell commands
being executed. This procedure was replaced with a single script that
returns all the files changed in all commits, with their appropriate
statuses. This way, the commit collection is 10x faster.

Other relationships and processing tasks are made with pandas, which
also saves some time.